Put extension in different file
