+++ Shared/Features/FreakDetails/View/FreakDetailsView.swift
@@ -1,17 +1,5 @@
 import SwiftUI
 
-// Get back swipe left gesture after hiding back button
-extension UINavigationController: UIGestureRecognizerDelegate {
-    override open func viewDidLoad() {
-        super.viewDidLoad()
-        interactivePopGestureRecognizer?.delegate = self
-    }
-
-    public func gestureRecognizerShouldBegin(_: UIGestureRecognizer) -> Bool {
-        viewControllers.count > 1
-    }
-}
-
 struct FreakDetailsView: View {
     @Environment(\.presentationMode) var presentationMode
     var viewModel: FreakDetailsViewModel
