+++ Shared/Utils/Extensions/UINavigationController+Extensions.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+// Get back swipe left gesture after hiding back button
+extension UINavigationController: UIGestureRecognizerDelegate {
+    override open func viewDidLoad() {
+        super.viewDidLoad()
+        interactivePopGestureRecognizer?.delegate = self
+    }
+
+    public func gestureRecognizerShouldBegin(_: UIGestureRecognizer) -> Bool {
+        viewControllers.count > 1
+    }
+}
